move make_metric and EvaluationMetric, add MetricValue
