+++ mlflow/metrics/__init__.py
@@ -1,0 +1,11 @@
+from mlflow.metrics.base import (
+    EvaluationMetric,
+    MetricValue,
+    make_metric,
+)
+
+__all__ = [
+    "EvaluationMetric",
+    "MetricValue",
+    "make_metric",
+]
