+++ mlflow/models/evaluation/__init__.py
@@ -1,12 +1,10 @@
 from mlflow.models.evaluation.base import (
     EvaluationArtifact,
     EvaluationDataset,
-    EvaluationMetric,
     EvaluationResult,
     ModelEvaluator,
     evaluate,
     list_evaluators,
-    make_metric,
 )
 from mlflow.models.evaluation.validation import MetricThreshold
 
@@ -14,9 +12,7 @@ __all__ = [
     "ModelEvaluator",
     "EvaluationDataset",
     "EvaluationResult",
-    "EvaluationMetric",
     "EvaluationArtifact",
-    "make_metric",
     "evaluate",
     "list_evaluators",
     "MetricThreshold",
